refactor(api): move create certification center controller
in src/organizational-entities
